feat: removed password prompt for encrypted backup and use vault pass
