+++ VultisigApp/VultisigApp/Localizables/en.lproj/Localizable.strings
@@ -808,3 +808,4 @@
 "referralCodeCopied" = "Referral code copied";
 "editFriendsReferral" = "Edit Friends Referral";
 "vaultSelected" = "Vault selected";
+"useMyVaultPassword" = "Use My Vault Password";
