--- VultisigApp/VultisigApp/iOS/View/PasswordBackupOptionsView+iOS.swift
@@ -1,36 +1,0 @@
-//
-//  PasswordBackupOptionsView+iOS.swift
-//  VultisigApp
-//
-//  Created by Amol Kumar on 2025-03-21.
-//
-
-#if os(iOS)
-import SwiftUI
-
-extension PasswordBackupOptionsView {
-    var content: some View {
-        VStack(spacing: 36) {
-            icon
-            textContent
-            buttons
-        }
-        .padding(24)
-    }
-    
-    @ViewBuilder
-    var withoutPasswordButton: some View {
-        if let fileURL = backupViewModel.encryptedFileURLWithoutPassword {
-            PrimaryButton(title: "backupWithoutPassword") {
-                showSkipShareSheet = true
-            }
-            .shareSheet(isPresented: $showSkipShareSheet, activityItems: [fileURL])  { didSave in
-                if didSave {
-                    fileSaved()
-                    dismissView()
-                }
-            }
-        }
-    }
-}
-#endif
