--- VultisigApp/VultisigApp/macOS/View/PasswordBackupOptionsView+macOS.swift
@@ -1,55 +1,0 @@
-//
-//  PasswordBackupOptionsView+macOS.swift
-//  VultisigApp
-//
-//  Created by Amol Kumar on 2025-03-21.
-//
-
-#if os(macOS)
-import SwiftUI
-
-extension PasswordBackupOptionsView {
-    var content: some View {
-        VStack(spacing: 36) {
-            header
-            Spacer()
-            
-            VStack(spacing: 36) {
-                icon
-                textContent
-                buttons
-            }
-            .padding(24)
-            
-            Spacer()
-        }
-    }
-    
-    var header: some View {
-        GeneralMacHeader(title: "")
-    }
-    
-    var withoutPasswordButton: some View {
-        PrimaryButton(title: "backupWithoutPassword") {
-            showSkipShareSheet = true
-        }
-        .fileExporter(
-            isPresented: $showSkipShareSheet,
-            document: EncryptedDataFile(url: backupViewModel.encryptedFileURLWithoutPassword),
-            contentType: .data,
-            defaultFilename: "\(vault.getExportName())"
-        ) { result in
-            switch result {
-            case .success(let url):
-                print("File saved to: \(url)")
-                fileSaved()
-                dismissView()
-            case .failure(let error):
-                print("Error saving file: \(error.localizedDescription)")
-                backupViewModel.alertTitle = "errorSavingFile"
-                backupViewModel.showAlert = true
-            }
-        }
-    }
-}
-#endif
